04._basic_servers/data_format_translation_servers/python/pyproject.toml
[tool.poetry]
name = "python"
version = "0.1.0"
description = ""
authors = ["madsWesth <72575386+madsWesth@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0"
xmltodict = "^0.13.0"
fastapi = "^0.92.0"
uvicorn = "^0.20.0"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
